+++ src/ts/modtypes/spectraVoice.ts
@@ -37,6 +37,11 @@ export namespace SpectraVoice {
     Overtones2 = 11,
     Overtones3 = 12,
     Overtones4 = 13,
+    Overtones1Plus = 14,
+    Overtones2Plus = 15,
+    Overtones3Plus = 16,
+    Overtones4Plus = 17,
+    Metal = 18,
   }
   export enum CtlNum {
     Volume = 1,